+++ src/ReplicateFilesystemAdapter.php
@@ -5,8 +5,16 @@ declare(strict_types=1);
 namespace Ajgl\Flysystem\Replicate;
 
 use League\Flysystem\FilesystemAdapter;
+use League\Flysystem\UrlGeneration\PublicUrlGenerator;
 
-final class ReplicateFilesystemAdapter implements FilesystemAdapter
-{
-    use ReplicateFilesystemAdapterTrait;
+if (interface_exists(PublicUrlGenerator::class)) {
+    final class ReplicateFilesystemAdapter implements FilesystemAdapter, PublicUrlGenerator
+    {
+        use ReplicateFilesystemAdapterTrait;
+    }
+} else {
+    final class ReplicateFilesystemAdapter implements FilesystemAdapter
+    {
+        use ReplicateFilesystemAdapterTrait;
+    }
 }